apis/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
cron-descriptor==1.4.0
croniter==2.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
orjson==3.9.12
psycopg2-binary==2.9.9
psycopg2
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
SQLAlchemy-Utils==0.38.3
starlette==0.35.1
typing-extensions==4.9.0
ujson==5.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
python-dotenv
